feat: multiplicative blood drinker (#54)
* feat: make blood drinker/thirst multiplicative bonuses

- Blood Drinker/Thirst are now multiplicative spell bonuses
- For missile weapons, ammo is now the buff target for these spells.

* fix: crit rate armor mod fixed

+++ Source/ACE.Server/WorldObjects/Managers/EnchantmentManagerWithCaching.cs
@@ -376,12 +376,12 @@ namespace ACE.Server.WorldObjects.Managers
         }
 
 
-        private int? damageBonusCache;
+        private float? damageBonusCache;
 
         /// <summary>
         /// Returns the weapon damage modifier, ie. Blood Drinker
         /// </summary>
-        public override int GetDamageBonus()
+        public override float GetDamageBonus()
         {
             if (damageBonusCache.HasValue)
                 return damageBonusCache.Value;
